node: Update npm dependencies
`npm update`, plus updates for uuid, node-gyp, and rimraf,
and removing now-unused @types/uuid and @types/long.

+++ node/ts/index.ts
@@ -831,7 +831,7 @@ export class SenderKeyDistributionMessage {
   ): Promise<SenderKeyDistributionMessage> {
     const handle = await Native.SenderKeyDistributionMessage_Create(
       sender,
-      Buffer.from(uuid.parse(distributionId) as Uint8Array),
+      Buffer.from(uuid.parse(distributionId)),
       store
     );
     return new SenderKeyDistributionMessage(handle);
@@ -848,7 +848,7 @@ export class SenderKeyDistributionMessage {
     return new SenderKeyDistributionMessage(
       Native.SenderKeyDistributionMessage_New(
         messageVersion,
-        Buffer.from(uuid.parse(distributionId) as Uint8Array),
+        Buffer.from(uuid.parse(distributionId)),
         chainId,
         iteration,
         chainKey,
@@ -912,7 +912,7 @@ export class SenderKeyMessage {
     return new SenderKeyMessage(
       Native.SenderKeyMessage_New(
         messageVersion,
-        Buffer.from(uuid.parse(distributionId) as Uint8Array),
+        Buffer.from(uuid.parse(distributionId)),
         chainId,
         iteration,
         ciphertext,
@@ -1214,7 +1214,7 @@ export async function groupEncrypt(
   return CiphertextMessage._fromNativeHandle(
     await Native.GroupCipher_EncryptMessage(
       sender,
-      Buffer.from(uuid.parse(distributionId) as Uint8Array),
+      Buffer.from(uuid.parse(distributionId)),
       message,
       store
     )
